requirements.txt
# KLIPAURA — API + worker + ugc_pipeline (core_v1)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0
boto3>=1.34.0
schedule>=1.2.0
redis>=5.0.0
requests
beautifulsoup4>=4.12.0
groq
elevenlabs
moviepy
pillow
opencv-python-headless
numpy
ffmpeg-python
httpx
sqlalchemy>=2.0
alembic>=1.13.0
psycopg[binary]>=3.1.0
